fix(maz-ui): MazPhoneNumberInput - use provide/inject for data to not sharing state between instances - issue #1024
